entity/address/tag cleanup

+++ server/src/handlers/v1/tags/delete.rs
@@ -14,9 +14,17 @@ pub async fn handler(
 	State(app): State<Arc<App>>,
 	Path(tag_id): Path<String>,
 ) -> ServerResult<StatusCode> {
-	if Tag::delete_by_id(app.db(), &tag_id).await? {
-		Ok(StatusCode::NO_CONTENT)
-	} else {
-		Err(ServerError::NotFound)
+	if let Some(tag) = Tag::get_by_id(app.db(), &tag_id).await? {
+		// if locked (@TODO and "sanctions" mode is on), don't allow deleting
+		if tag.is_locked {
+			return Err(ServerError::BadRequest { reason: "object is locked".to_string() });
+		}
+
+		// delete
+		if Tag::delete_by_id(app.db(), &tag_id).await? {
+			return Ok(StatusCode::NO_CONTENT);
+		}
 	}
+
+	Err(ServerError::NotFound)
 }
